pyproject.toml
[project]
name = "shop-profit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dbt-duckdb>=1.10.1",
    "duckdb>=1.4.4",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
